+++ sphinxcontrib/jsonschema.py
@@ -33,6 +33,7 @@ class JSONSchemaDirective(Directive):
         'pointer': directives.unchanged,
         'nocrossref': directives.flag,
         'addtargets': directives.flag,
+        'externallinks': directives.unchanged,
     }
     # Add a rollup option here
 
@@ -42,6 +43,7 @@ class JSONSchemaDirective(Directive):
     include_used = set()
     collapse = []
     collapse_used = set()
+    external_links = {}
 
     def run(self):
         include = self.options.get('include')
@@ -52,6 +54,10 @@ class JSONSchemaDirective(Directive):
         if collapse:
             self.collapse = collapse.split(',')
             self.collapse_used = set()
+
+        external_links = self.options.get('externallinks')
+        if external_links:
+            self.external_links = json.loads(external_links)
 
         env = self.state.document.settings.env
         try:
@@ -143,7 +149,16 @@ class JSONSchemaDirective(Directive):
         row += self.cell(prop.title)
         if prop.description:
             cell = self.cell(prop.description or '', morecols=3)
-            if 'nocrossref' not in self.options:
+            if prop.name in self.external_links:
+                reference = nodes.reference(
+                    '',
+                    '',
+                    nodes.Text(self.external_links[prop.name]['text']),
+                    internal=False,
+                    refuri=self.external_links[prop.name]['url'],
+                    anchorname='')
+                cell += nodes.paragraph('', nodes.Text('\n\nSee '), reference)
+            elif 'nocrossref' not in self.options:
                 ref = None
                 if hasattr(prop.attributes, '__reference__'):
                     ref = prop.attributes.__reference__['$ref']